requirements.txt
python-telegram-bot==20.7
openpyxl
pandas
whisper
torch
ffmpeg-python
